Implement new setTimeout and setInterval functions
Javascript's built-in timing functions stop working when the application
is in background state. This causes race conditions. The new
implementation doesn't depend on a ticking clock but on a fixed due date
that is computed when the timer is set. This allows the timer to still
work even if it has gone to sleep for some time.

Issue #1227